feat: criacao do metodo de retorno do cart e criacao da cartindex

+++ GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -8,6 +8,6 @@
 
         public string CouponCode { get; set; }
 
-        public double PurchaseAmount { get; set; } = 0;
+        public decimal PurchaseAmount { get; set; } = 0;
     }
 }
